Create Add_Abundance_Polygons.R
This function takes a table of abundances found on each sample (rows), e.g. relative abundances of different microbial groups (columns), and gets the global abundance of each microbial group among all samples. Then, it draws a rectangle on the corresponding coordinates of the multivariate analysis (NMDS or PCA) of proportional size to its global abundance in experiment analyzed.