Create Get_Overrepresented_Groups.R
This is an additional function which works alone but can be a complement to Add_Abundance_Polygons. From an abundance table (e.g. relative abundances of microbial groups in  a set of samples), it allows to cluster samples in different groups and retrieved which columns (microbial groups) show abundances only in a specific group of samples or when the abundance in a group is x-fold higher than the others.