Create PCAvars.R
This is an addtional function to Principal Components Analysis (PCA) made by prcomp function from packages stats. It adds to the pca analysis relative contribution of each variable to the 2D representation of the PCA and it can also add these variables to the graphical representation if it is required.